Calculate predicted abundance

Calculates the predicted abundance by multiplying the density prediction values by prediction polygon areas

model_abundance(x, dens.idx, sum.abund = TRUE)

Arguments

x
object of class sf; SDM with density predictions. Must have a valid crs code

dens.idx
name or index of column(s) in x with density predictions. Can be a character vector (column names) or numeric vector (column indices)

sum.abund
logical; whether or not to sum all of the predicted abundances

Value

If sum.abund == TRUE, then a vector of the same length as dens.idx representing the predicted abundance for the density values in each column.

If sum.abund == FALSE and the length of dens.idx is 1, then a numeric vector with the predicted abundance of each prediction polygon of x.

If sum.abund == FALSE and the length of dens.idx is greater than 1, then a data frame with length(dens.idx) columns of the predicted abundance of prediction polygons

Details

Multiplies the values in the specified column(s) (i.e. the density predictions) by the area in square kilometers of their corresponding prediction polygon. The area of each prediction polygon is calculated using st_area from geos_measures. x must have a valid crs code to calculate area for these abundance calculations.

Examples

model_abundance(preds.1, "Density")

#>  Density 
#> 113746.5 

model_abundance(preds.1, c(1, 1))

#> [1] 113746.5 113746.5

model_abundance(preds.1, c(1, 1), FALSE)

#>         1.abund     1.abund
#> 1     29.694692   29.694692
#> 2      9.206602    9.206602
#> 3     21.603062   21.603062
#> 4      9.503845    9.503845
#> 5     11.984480   11.984480
#> 6     15.298464   15.298464
#> 7      7.686818    7.686818
#> 8      9.459501    9.459501
#> 9     13.948896   13.948896
#> 10    48.275487   48.275487
#> 11    81.225826   81.225826
#> 12    37.756303   37.756303
#> 13    21.109040   21.109040
#> 14    29.351664   29.351664
#> 15   308.464960  308.464960
#> 16   510.867257  510.867257
#> 17   557.909056  557.909056
#> 18   653.772706  653.772706
#> 19   438.223086  438.223086
#> 20   142.703556  142.703556
#> 21   392.833681  392.833681
#> 22   496.217782  496.217782
#> 23   639.299448  639.299448
#> 24   669.838598  669.838598
#> 25   251.321053  251.321053
#> 26    49.088068   49.088068
#> 27    51.319426   51.319426
#> 28    21.116404   21.116404
#> 29    18.524162   18.524162
#> 30    14.258505   14.258505
#> 31    14.655641   14.655641
#> 32    33.422614   33.422614
#> 33    25.665085   25.665085
#> 34    59.504226   59.504226
#> 35   102.071523  102.071523
#> 36   161.895465  161.895465
#> 37    31.546858   31.546858
#> 38    60.752537   60.752537
#> 39   439.713436  439.713436
#> 40   978.177906  978.177906
#> 41   969.672751  969.672751
#> 42   958.938936  958.938936
#> 43  1075.726668 1075.726668
#> 44   526.629908  526.629908
#> 45   491.766313  491.766313
#> 46   500.445405  500.445405
#> 47  1177.495658 1177.495658
#> 48  1265.753353 1265.753353
#> 49   831.075711  831.075711
#> 50   723.813931  723.813931
#> 51    48.393730   48.393730
#> 52    22.281642   22.281642
#> 53    13.499157   13.499157
#> 54    13.371317   13.371317
#> 55     7.988059    7.988059
#> 56    13.294254   13.294254
#> 57     4.606512    4.606512
#> 58    41.981277   41.981277
#> 59    53.878155   53.878155
#> 60   134.639241  134.639241
#> 61   277.829997  277.829997
#> 62    62.100636   62.100636
#> 63   766.650992  766.650992
#> 64  1021.706795 1021.706795
#> 65  1099.304373 1099.304373
#> 66  1318.499607 1318.499607
#> 67  1263.807308 1263.807308
#> 68  1025.154113 1025.154113
#> 69  1074.339275 1074.339275
#> 70   567.855726  567.855726
#> 71   462.824097  462.824097
#> 72   846.060947  846.060947
#> 73  1007.214840 1007.214840
#> 74  1394.181764 1394.181764
#> 75           NA          NA
#> 76    49.184978   49.184978
#> 77    14.903474   14.903474
#> 78     5.145944    5.145944
#> 79     4.748368    4.748368
#> 80    16.461609   16.461609
#> 81    22.136921   22.136921
#> 82     6.197428    6.197428
#> 83    63.791711   63.791711
#> 84   183.294326  183.294326
#> 85   333.510671  333.510671
#> 86    75.932207   75.932207
#> 87   660.341472  660.341472
#> 88  1136.459432 1136.459432
#> 89  1257.229465 1257.229465
#> 90  1176.554255 1176.554255
#> 91  1372.678352 1372.678352
#> 92  1050.843585 1050.843585
#> 93  1099.143060 1099.143060
#> 94   619.240769  619.240769
#> 95   644.439989  644.439989
#> 96  1021.023357 1021.023357
#> 97  1004.018207 1004.018207
#> 98  1198.913902 1198.913902
#> 99           NA          NA
#> 100          NA          NA
#> 101   99.972107   99.972107
#> 102   20.625530   20.625530
#> 103    2.822829    2.822829
#> 104    2.985765    2.985765
#> 105    8.512578    8.512578
#> 106   28.688387   28.688387
#> 107   68.452453   68.452453
#> 108  160.779234  160.779234
#> 109   49.612508   49.612508
#> 110  232.578641  232.578641
#> 111  351.300769  351.300769
#> 112 1373.124865 1373.124865
#> 113 1252.396670 1252.396670
#> 114  992.550731  992.550731
#> 115 1105.298474 1105.298474
#> 116  835.442557  835.442557
#> 117 1207.809237 1207.809237
#> 118  799.810069  799.810069
#> 119  601.169216  601.169216
#> 120 1131.991484 1131.991484
#> 121 1251.497481 1251.497481
#> 122 1364.304652 1364.304652
#> 123 1337.465896 1337.465896
#> 124          NA          NA
#> 125          NA          NA
#> 126   93.004972   93.004972
#> 127   31.498775   31.498775
#> 128    1.277841    1.277841
#> 129    1.441917    1.441917
#> 130   11.099438   11.099438
#> 131   58.641385   58.641385
#> 132  125.364892  125.364892
#> 133   76.535644   76.535644
#> 134  106.398113  106.398113
#> 135  196.264649  196.264649
#> 136 1025.919442 1025.919442
#> 137 1202.063302 1202.063302
#> 138 1065.754690 1065.754690
#> 139 1053.229294 1053.229294
#> 140          NA          NA
#> 141  625.290331  625.290331
#> 142  794.004610  794.004610
#> 143 1136.470132 1136.470132
#> 144 1054.589932 1054.589932
#> 145 1198.324780 1198.324780
#> 146 1422.711401 1422.711401
#> 147 1412.311820 1412.311820
#> 148 1286.593577 1286.593577
#> 149          NA          NA
#> 150          NA          NA
#> 151   73.897729   73.897729
#> 152   40.592099   40.592099
#> 153    3.910365    3.910365
#> 154    4.824639    4.824639
#> 155   29.499293   29.499293
#> 156   29.079375   29.079375
#> 157   78.968933   78.968933
#> 158  184.642093  184.642093
#> 159  149.913266  149.913266
#> 160  220.604706  220.604706
#> 161 1277.834996 1277.834996
#> 162  862.142192  862.142192
#> 163  667.758821  667.758821
#> 164  703.701487  703.701487
#> 165  537.992485  537.992485
#> 166  807.539937  807.539937
#> 167 1128.780362 1128.780362
#> 168 1213.255402 1213.255402
#> 169 1323.162024 1323.162024
#> 170 1388.235961 1388.235961
#> 171 1428.484834 1428.484834
#> 172          NA          NA
#> 173          NA          NA
#> 174          NA          NA
#> 175          NA          NA
#> 176  118.785252  118.785252
#> 177   18.119765   18.119765
#> 178   15.498591   15.498591
#> 179   14.560943   14.560943
#> 180   35.850678   35.850678
#> 181  110.314117  110.314117
#> 182  124.955846  124.955846
#> 183   28.664450   28.664450
#> 184   31.200054   31.200054
#> 185   92.061410   92.061410
#> 186  258.922071  258.922071
#> 187  684.652610  684.652610
#> 188  993.198919  993.198919
#> 189  639.481219  639.481219
#> 190  658.302382  658.302382
#> 191 1064.816721 1064.816721
#> 192 1276.720939 1276.720939
#> 193 1360.841639 1360.841639
#> 194 1182.027000 1182.027000
#> 195          NA          NA
#> 196          NA          NA
#> 197          NA          NA
#> 198          NA          NA
#> 199          NA          NA
#> 200          NA          NA
#> 201  123.650698  123.650698
#> 202  145.116188  145.116188
#> 203   48.045987   48.045987
#> 204   60.359197   60.359197
#> 205   81.507816   81.507816
#> 206  108.451318  108.451318
#> 207   63.776426   63.776426
#> 208   20.418998   20.418998
#> 209   33.355505   33.355505
#> 210   88.573681   88.573681
#> 211  298.914818  298.914818
#> 212          NA          NA
#> 213 1398.925722 1398.925722
#> 214          NA          NA
#> 215 1200.994960 1200.994960
#> 216 1145.872600 1145.872600
#> 217  899.924983  899.924983
#> 218  738.948164  738.948164
#> 219 1332.446230 1332.446230
#> 220          NA          NA
#> 221          NA          NA
#> 222          NA          NA
#> 223          NA          NA
#> 224          NA          NA
#> 225          NA          NA
#> 226  148.411093  148.411093
#> 227  277.487810  277.487810
#> 228   97.262211   97.262211
#> 229   90.056296   90.056296
#> 230  139.666135  139.666135
#> 231  102.826127  102.826127
#> 232   18.268087   18.268087
#> 233   44.717649   44.717649
#> 234   91.994242   91.994242
#> 235  579.301165  579.301165
#> 236  829.460525  829.460525
#> 237 1192.706115 1192.706115
#> 238 1302.150295 1302.150295
#> 239 1383.350878 1383.350878
#> 240 1398.860516 1398.860516
#> 241          NA          NA
#> 242          NA          NA
#> 243          NA          NA
#> 244          NA          NA
#> 245          NA          NA
#> 246          NA          NA
#> 247          NA          NA
#> 248          NA          NA
#> 249          NA          NA
#> 250          NA          NA
#> 251  219.828905  219.828905
#> 252  253.159963  253.159963
#> 253  161.398237  161.398237
#> 254  233.493212  233.493212
#> 255  164.022600  164.022600
#> 256   30.894574   30.894574
#> 257   83.088135   83.088135
#> 258  108.862149  108.862149
#> 259  229.800847  229.800847
#> 260  356.712378  356.712378
#> 261          NA          NA
#> 262          NA          NA
#> 263          NA          NA
#> 264          NA          NA
#> 265          NA          NA
#> 266          NA          NA
#> 267          NA          NA
#> 268          NA          NA
#> 269          NA          NA
#> 270          NA          NA
#> 271          NA          NA
#> 272          NA          NA
#> 273          NA          NA
#> 274          NA          NA
#> 275          NA          NA
#> 276  221.718576  221.718576
#> 277  332.692849  332.692849
#> 278  213.463656  213.463656
#> 279  243.120945  243.120945
#> 280   63.730533   63.730533
#> 281  147.329529  147.329529
#> 282  700.108996  700.108996
#> 283  972.574187  972.574187
#> 284  534.244977  534.244977
#> 285  632.502879  632.502879
#> 286          NA          NA
#> 287          NA          NA
#> 288          NA          NA
#> 289          NA          NA
#> 290          NA          NA
#> 291          NA          NA
#> 292          NA          NA
#> 293          NA          NA
#> 294          NA          NA
#> 295          NA          NA
#> 296          NA          NA
#> 297          NA          NA
#> 298          NA          NA
#> 299          NA          NA
#> 300          NA          NA
#> 301   95.553662   95.553662
#> 302  126.688413  126.688413
#> 303   52.805044   52.805044
#> 304   41.873327   41.873327
#> 305    9.222216    9.222216
#> 306  120.766492  120.766492
#> 307  241.896136  241.896136
#> 308  477.623067  477.623067
#> 309  307.530333  307.530333
#> 310  213.983649  213.983649
#> 311          NA          NA
#> 312          NA          NA
#> 313          NA          NA
#> 314          NA          NA
#> 315          NA          NA
#> 316          NA          NA
#> 317          NA          NA
#> 318          NA          NA
#> 319          NA          NA
#> 320          NA          NA
#> 321          NA          NA
#> 322          NA          NA
#> 323          NA          NA
#> 324          NA          NA
#> 325          NA          NA