Overlay SDM predictions onto base geometry

Overlay specified SDM predictions that meet the percent overlap threshold requirement onto base geometry

Usage

overlay_sdm(base.geom, sdm, sdm.idx, overlap.perc)

Arguments

base.geom

object of class sfc; base geometry

sdm

object of class sf; original SDM predictions

sdm.idx

names or indices of column(s) with data to be overlaid

overlap.perc

numeric; percent overlap threshold, i.e. percentage of each base geometry polygon must overlap with SDM prediction polygons for overlaid density value to be calculated and not set as NA

Value

Object of class sf with the geometry of base.geom and the data in the sdm.idx columns of sdm overlaid onto that geometry. Note that this means all columns of sdm not in sdm.idx will not be in the returned object. Because the data are considered spatially intensive, the agr attribute will be set as 'constant' for all columns in the returned object

Details

See the eSDM GUI manual for specifics about the overlay process. This process is equivalent to areal interpolation (Goodchild and Lam 1980), where base.geom is the target, sdm is the source, and the data specified by sdm.idx are spatially intensive.

Note that overlay_sdm removes rows in sdm that have NA values in the first column specified in sdm.idx (i.e. sdm.idx[1]), before the overlay. Thus, for valid overlay results, all columns of sdm specified in sdm.idx must either have NA values in the same rows or contain only NAs.

References

Goodchild, M.F. & Lam, N.S.-N. (1980) Areal interpolation: a variant of the traditional spatial problem. Geo-Processing, 1, 297-312.

Examples

pol1.geom <- sf::st_sfc(
  sf::st_polygon(list(rbind(c(1,1), c(3,1), c(3,3), c(1,3), c(1,1)))),
  crs = 4326
)
pol2.geom <- sf::st_sfc(
  sf::st_polygon(list(rbind(c(0,0), c(2,0), c(2,2), c(0,2), c(0,0)))),
  crs = 4326
)
pol2.sf <- sf::st_sf(data.frame(Dens = 0.5), geometry = pol2.geom, crs = 4326)

overlay_sdm(pol1.geom, pol2.sf, 1, 25)
#> Simple feature collection with 1 feature and 1 field
#> Attribute-geometry relationship: 1 constant, 0 aggregate, 0 identity
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 1 ymin: 1 xmax: 3 ymax: 3
#> Geodetic CRS:  WGS 84
#>   Dens                       geometry
#> 1  0.5 POLYGON ((1 1, 3 1, 3 3, 1 ...

# Output 'Dens' value is NA because of higher overlap.perc value
overlay_sdm(pol1.geom, pol2.sf, 1, 50)
#> Simple feature collection with 1 feature and 1 field
#> Attribute-geometry relationship: 1 constant, 0 aggregate, 0 identity
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 1 ymin: 1 xmax: 3 ymax: 3
#> Geodetic CRS:  WGS 84
#>   Dens                       geometry
#> 1   NA POLYGON ((1 1, 3 1, 3 3, 1 ...

# These examples take longer to run
overlay_sdm(sf::st_geometry(preds.1), preds.2, 1, 50)
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> Warning: The output of st_intersection(sdm, base.geom) was invalid - this was corrected via sf::st_make_valid(st_intersection(sdm, base.geom))
#> Simple feature collection with 325 features and 1 field
#> Attribute-geometry relationship: 1 constant, 0 aggregate, 0 identity
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -123 ymin: 32 xmax: -117 ymax: 35
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> Geodetic CRS:  WGS 84
#> First 10 features:
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#>         Density                       geometry
#> 1  0.0007590172 POLYGON ((-122.875 32.125, ...
#> 2  0.0006972454 POLYGON ((-122.625 32.125, ...
#> 3  0.0010751592 POLYGON ((-122.375 32.125, ...
#> 4  0.0018757942 POLYGON ((-122.125 32.125, ...
#> 5  0.0022005824 POLYGON ((-121.875 32.125, ...
#> 6  0.0023619783 POLYGON ((-121.625 32.125, ...
#> 7  0.0021131121 POLYGON ((-121.375 32.125, ...
#> 8  0.0019934417 POLYGON ((-121.125 32.125, ...
#> 9  0.0020992076 POLYGON ((-120.875 32.125, ...
#> 10 0.0020969021 POLYGON ((-120.625 32.125, ...
overlay_sdm(sf::st_geometry(preds.2), preds.1, "Density", 50)
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> Warning: The output of st_intersection(sdm, base.geom) was invalid - this was corrected via sf::st_make_valid(st_intersection(sdm, base.geom))
#> Simple feature collection with 1891 features and 1 field
#> Attribute-geometry relationship: 1 constant, 0 aggregate, 0 identity
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -123 ymin: 32 xmax: -117 ymax: 35
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> Geodetic CRS:  WGS 84
#> First 10 features:
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#> old-style crs object detected; please recreate object with a recent sf::st_crs()
#>       Density                       geometry
#> 1  0.18149697 POLYGON ((-122.95 32.05, -1...
#> 2  0.14315865 POLYGON ((-122.85 32.05, -1...
#> 3  0.02813584 POLYGON ((-122.75 32.05, -1...
#> 4  0.02813584 POLYGON ((-122.65 32.05, -1...
#> 5  0.05654697 POLYGON ((-122.55 32.05, -1...
#> 6  0.06602004 POLYGON ((-122.45 32.05, -1...
#> 7  0.05677412 POLYGON ((-122.35 32.05, -1...
#> 8  0.02904423 POLYGON ((-122.25 32.05, -1...
#> 9  0.02904423 POLYGON ((-122.15 32.05, -1...
#> 10 0.03472953 POLYGON ((-122.05 32.05, -1...