Rescale SDM predictions

Rescale SDM predictions and (if applicable) associated uncertainties

ensemble_rescale(x, x.idx, y, y.abund = NULL, x.var.idx = NULL)

Arguments

x
object of class sf

x.idx
vector of column names or column indices; indicates columns in x with prediction values that will be rescaled

y
rescaling method; must be either "abundance" or "sumto1". See 'Details' section for descriptions of the rescaling methods

y.abund
numeric value; ignored if y is not "abundance"

x.var.idx
vector of column names or column indices; indicates columns in x with variance values that will be rescaled. If x.var.idx is specified, it must be the same length as x.idx. Use x.var.idx = NULL (the default) if none of the predictions have associated uncertainty values; see the 'Details' section for more information

Value

The sf object x with the columns specified by x.idx and x.var.idx rescaled. The agr attributes of x will be conserved

Details

ensemble_rescale is intended to be used after overlaying predictions with overlay_sdm and before creating ensembles with ensemble_create. The provided rescaling methods are:

'abundance' - Rescale the density values so that the predicted abundance is y.abund

'sumto1' - Rescale the density values so their sum is 1

SDM uncertainty values must be rescaled differently than the prediction values. Columns specified in x.var.idx must contain variance values. These values will be rescaled using the formula var(c * x) = c^2 * var(x), where c is the rescaling factor for the associated predictions.

If x.var.idx is not NULL, then the function assumes x.var.idx[1] contains the variance values associated with the predictions in x.idx[1], x.var.idx[2] contains the variance values associated with the predictions in x.idx[2], etc. Use NA in x.var.idx to indicate a set of predictions that does not have associated uncertainty values (e.g., x.var.idx = c(4, NA, 5))

Examples

ensemble_rescale(preds.1, c("Density", "Density2"), "abundance", 50)

#> Linking to GEOS 3.6.1, GDAL 2.2.3, PROJ 4.9.3

#> Simple feature collection with 325 features and 4 fields
#> Attribute-geometry relationship: 4 constant, 0 aggregate, 0 identity
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: -123 ymin: 32 xmax: -117 ymax: 35
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs
#> First 10 features:
#>         Density     Density2         Var1         Var2
#> 1  7.978133e-05 1.529755e-04 0.0006178977 6.782811e-03
#> 2  1.236778e-05 2.113414e-04 0.0041072496 7.100138e-03
#> 3  2.902069e-05 1.013807e-04 0.0060756461 2.152463e-03
#> 4  1.276708e-05 5.109601e-04 0.0030639979 7.415022e-03
#> 5  1.609947e-05 9.858766e-06 0.0009543514 2.318842e-03
#> 6  2.055134e-05 3.325207e-04 0.0081108828 1.309096e-05
#> 7  1.032616e-05 3.909524e-04 0.0068363941 2.566385e-04
#> 8  1.270751e-05 5.118330e-04 0.0047392882 7.242310e-03
#> 9  1.873839e-05 2.753378e-05 0.0038668358 5.157556e-03
#> 10 6.485135e-05 3.566653e-04 0.0071928494 7.077782e-03
#>                          geometry
#> 1  POLYGON ((-122.875 32.125, ...
#> 2  POLYGON ((-122.625 32.125, ...
#> 3  POLYGON ((-122.375 32.125, ...
#> 4  POLYGON ((-122.125 32.125, ...
#> 5  POLYGON ((-121.875 32.125, ...
#> 6  POLYGON ((-121.625 32.125, ...
#> 7  POLYGON ((-121.375 32.125, ...
#> 8  POLYGON ((-121.125 32.125, ...
#> 9  POLYGON ((-120.875 32.125, ...
#> 10 POLYGON ((-120.625 32.125, ...

ensemble_rescale(preds.1, c(1, 2), "sumto1")

#> Simple feature collection with 325 features and 4 fields
#> Attribute-geometry relationship: 4 constant, 0 aggregate, 0 identity
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: -123 ymin: 32 xmax: -117 ymax: 35
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs
#> First 10 features:
#>         Density     Density2         Var1         Var2
#> 1  0.0009493594 0.0017608119 0.0006178977 6.782811e-03
#> 2  0.0001471706 0.0024326274 0.0041072496 7.100138e-03
#> 3  0.0003453322 0.0011669347 0.0060756461 2.152463e-03
#> 4  0.0001519221 0.0058813643 0.0030639979 7.415022e-03
#> 5  0.0001915759 0.0001134785 0.0009543514 2.318842e-03
#> 6  0.0002445511 0.0038274525 0.0081108828 1.309096e-05
#> 7  0.0001228764 0.0045000253 0.0068363941 2.566385e-04
#> 8  0.0001512133 0.0058914121 0.0047392882 7.242310e-03
#> 9  0.0002229778 0.0003169253 0.0038668358 5.157556e-03
#> 10 0.0007716998 0.0041053668 0.0071928494 7.077782e-03
#>                          geometry
#> 1  POLYGON ((-122.875 32.125, ...
#> 2  POLYGON ((-122.625 32.125, ...
#> 3  POLYGON ((-122.375 32.125, ...
#> 4  POLYGON ((-122.125 32.125, ...
#> 5  POLYGON ((-121.875 32.125, ...
#> 6  POLYGON ((-121.625 32.125, ...
#> 7  POLYGON ((-121.375 32.125, ...
#> 8  POLYGON ((-121.125 32.125, ...
#> 9  POLYGON ((-120.875 32.125, ...
#> 10 POLYGON ((-120.625 32.125, ...

ensemble_rescale(
  preds.1, c("Density", "Density2"), "abundance", 100, c(3,4)
)

#> Simple feature collection with 325 features and 4 fields
#> Attribute-geometry relationship: 4 constant, 0 aggregate, 0 identity
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: -123 ymin: 32 xmax: -117 ymax: 35
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs
#> First 10 features:
#>         Density     Density2         Var1         Var2
#> 1  1.595627e-04 3.059510e-04 4.775737e-10 8.088049e-09
#> 2  2.473556e-05 4.226827e-04 3.174497e-09 8.466440e-09
#> 3  5.804137e-05 2.027615e-04 4.695872e-09 2.566669e-09
#> 4  2.553417e-05 1.021920e-03 2.368166e-09 8.841918e-09
#> 5  3.219894e-05 1.971753e-05 7.376190e-10 2.765064e-09
#> 6  4.110268e-05 6.650415e-04 6.268908e-09 1.561010e-11
#> 7  2.065233e-05 7.819048e-04 5.283855e-09 3.060243e-10
#> 8  2.541503e-05 1.023666e-03 3.663000e-09 8.635971e-09
#> 9  3.747677e-05 5.506756e-05 2.988681e-09 6.150041e-09
#> 10 1.297027e-04 7.133306e-04 5.559359e-09 8.439782e-09
#>                          geometry
#> 1  POLYGON ((-122.875 32.125, ...
#> 2  POLYGON ((-122.625 32.125, ...
#> 3  POLYGON ((-122.375 32.125, ...
#> 4  POLYGON ((-122.125 32.125, ...
#> 5  POLYGON ((-121.875 32.125, ...
#> 6  POLYGON ((-121.625 32.125, ...
#> 7  POLYGON ((-121.375 32.125, ...
#> 8  POLYGON ((-121.125 32.125, ...
#> 9  POLYGON ((-120.875 32.125, ...
#> 10 POLYGON ((-120.625 32.125, ...