Calculate predicted abundance

Calculates the predicted abundance by multiplying the density prediction values by prediction polygon areas

model_abundance(x, dens.idx, sum.abund = TRUE)

Arguments

x
object of class sf; SDM with density predictions. Must have a valid crs code

dens.idx
name or index of column(s) in x with density predictions. Can be a character vector (column names) or numeric vector (column indices)

sum.abund
logical; whether or not to sum all of the predicted abundances

Value

If sum.abund == TRUE, then a vector of the same length as dens.idx representing the predicted abundance for the density values in each column.

If sum.abund == FALSE and the length of dens.idx is 1, then a numeric vector with the predicted abundance of each prediction polygon of x.

If sum.abund == FALSE and the length of dens.idx is greater than 1, then a data frame with length(dens.idx) columns of the predicted abundance of prediction polygons

Details

Multiplies the values in the specified column(s) (i.e. the density predictions) by the area in square kilometers of their corresponding prediction polygon. The area of each prediction polygon is calculated using st_area from geos_measures. x must have a valid crs code to calculate area for these abundance calculations.

Examples

model_abundance(preds.1, "Density")

#>  Density 
#> 113746.5 

model_abundance(preds.1, c(1, 1))

#> [1] 113746.5 113746.5

model_abundance(preds.1, c(1, 1), FALSE)

#>         1.abund     1.abund
#> 1     29.694694   29.694694
#> 2      9.206602    9.206602
#> 3     21.603063   21.603063
#> 4      9.503845    9.503845
#> 5     11.984481   11.984481
#> 6     15.298465   15.298465
#> 7      7.686818    7.686818
#> 8      9.459501    9.459501
#> 9     13.948897   13.948897
#> 10    48.275490   48.275490
#> 11    81.225831   81.225831
#> 12    37.756306   37.756306
#> 13    21.109041   21.109041
#> 14    29.351666   29.351666
#> 15   308.464979  308.464979
#> 16   510.867289  510.867289
#> 17   557.909091  557.909091
#> 18   653.772747  653.772747
#> 19   438.223113  438.223113
#> 20   142.703565  142.703565
#> 21   392.833706  392.833706
#> 22   496.217813  496.217813
#> 23   639.299489  639.299489
#> 24   669.838640  669.838640
#> 25   251.321069  251.321069
#> 26    49.088071   49.088071
#> 27    51.319429   51.319429
#> 28    21.116405   21.116405
#> 29    18.524163   18.524163
#> 30    14.258506   14.258506
#> 31    14.655642   14.655642
#> 32    33.422616   33.422616
#> 33    25.665087   25.665087
#> 34    59.504230   59.504230
#> 35   102.071530  102.071530
#> 36   161.895474  161.895474
#> 37    31.546860   31.546860
#> 38    60.752540   60.752540
#> 39   439.713462  439.713462
#> 40   978.177964  978.177964
#> 41   969.672808  969.672808
#> 42   958.938993  958.938993
#> 43  1075.726731 1075.726731
#> 44   526.629940  526.629940
#> 45   491.766343  491.766343
#> 46   500.445435  500.445435
#> 47  1177.495728 1177.495728
#> 48  1265.753428 1265.753428
#> 49   831.075760  831.075760
#> 50   723.813974  723.813974
#> 51    48.393733   48.393733
#> 52    22.281644   22.281644
#> 53    13.499158   13.499158
#> 54    13.371318   13.371318
#> 55     7.988059    7.988059
#> 56    13.294255   13.294255
#> 57     4.606512    4.606512
#> 58    41.981280   41.981280
#> 59    53.878158   53.878158
#> 60   134.639249  134.639249
#> 61   277.830012  277.830012
#> 62    62.100639   62.100639
#> 63   766.651034  766.651034
#> 64  1021.706850 1021.706850
#> 65  1099.304433 1099.304433
#> 66  1318.499679 1318.499679
#> 67  1263.807376 1263.807376
#> 68  1025.154169 1025.154169
#> 69  1074.339333 1074.339333
#> 70   567.855757  567.855757
#> 71   462.824122  462.824122
#> 72   846.060993  846.060993
#> 73  1007.214895 1007.214895
#> 74  1394.181840 1394.181840
#> 75           NA          NA
#> 76    49.184981   49.184981
#> 77    14.903475   14.903475
#> 78     5.145944    5.145944
#> 79     4.748369    4.748369
#> 80    16.461610   16.461610
#> 81    22.136922   22.136922
#> 82     6.197428    6.197428
#> 83    63.791714   63.791714
#> 84   183.294335  183.294335
#> 85   333.510687  333.510687
#> 86    75.932211   75.932211
#> 87   660.341505  660.341505
#> 88  1136.459489 1136.459489
#> 89  1257.229528 1257.229528
#> 90  1176.554313 1176.554313
#> 91  1372.678420 1372.678420
#> 92  1050.843637 1050.843637
#> 93  1099.143114 1099.143114
#> 94   619.240800  619.240800
#> 95   644.440021  644.440021
#> 96  1021.023407 1021.023407
#> 97  1004.018257 1004.018257
#> 98  1198.913962 1198.913962
#> 99           NA          NA
#> 100          NA          NA
#> 101   99.972111   99.972111
#> 102   20.625531   20.625531
#> 103    2.822829    2.822829
#> 104    2.985766    2.985766
#> 105    8.512578    8.512578
#> 106   28.688388   28.688388
#> 107   68.452456   68.452456
#> 108  160.779241  160.779241
#> 109   49.612510   49.612510
#> 110  232.578652  232.578652
#> 111  351.300784  351.300784
#> 112 1373.124926 1373.124926
#> 113 1252.396726 1252.396726
#> 114  992.550776  992.550776
#> 115 1105.298524 1105.298524
#> 116  835.442595  835.442595
#> 117 1207.809291 1207.809291
#> 118  799.810105  799.810105
#> 119  601.169243  601.169243
#> 120 1131.991535 1131.991535
#> 121 1251.497538 1251.497538
#> 122 1364.304713 1364.304713
#> 123 1337.465957 1337.465957
#> 124          NA          NA
#> 125          NA          NA
#> 126   93.004976   93.004976
#> 127   31.498776   31.498776
#> 128    1.277841    1.277841
#> 129    1.441917    1.441917
#> 130   11.099439   11.099439
#> 131   58.641388   58.641388
#> 132  125.364897  125.364897
#> 133   76.535648   76.535648
#> 134  106.398117  106.398117
#> 135  196.264657  196.264657
#> 136 1025.919483 1025.919483
#> 137 1202.063351 1202.063351
#> 138 1065.754732 1065.754732
#> 139 1053.229336 1053.229336
#> 140          NA          NA
#> 141  625.290356  625.290356
#> 142  794.004642  794.004642
#> 143 1136.470178 1136.470178
#> 144 1054.589975 1054.589975
#> 145 1198.324828 1198.324828
#> 146 1422.711459 1422.711459
#> 147 1412.311876 1412.311876
#> 148 1286.593628 1286.593628
#> 149          NA          NA
#> 150          NA          NA
#> 151   73.897732   73.897732
#> 152   40.592100   40.592100
#> 153    3.910365    3.910365
#> 154    4.824639    4.824639
#> 155   29.499294   29.499294
#> 156   29.079376   29.079376
#> 157   78.968936   78.968936
#> 158  184.642099  184.642099
#> 159  149.913271  149.913271
#> 160  220.604714  220.604714
#> 161 1277.835041 1277.835041
#> 162  862.142223  862.142223
#> 163  667.758844  667.758844
#> 164  703.701512  703.701512
#> 165  537.992504  537.992504
#> 166  807.539965  807.539965
#> 167 1128.780402 1128.780402
#> 168 1213.255445 1213.255445
#> 169 1323.162071 1323.162071
#> 170 1388.236010 1388.236010
#> 171 1428.484884 1428.484884
#> 172          NA          NA
#> 173          NA          NA
#> 174          NA          NA
#> 175          NA          NA
#> 176  118.785256  118.785256
#> 177   18.119766   18.119766
#> 178   15.498592   15.498592
#> 179   14.560944   14.560944
#> 180   35.850679   35.850679
#> 181  110.314120  110.314120
#> 182  124.955850  124.955850
#> 183   28.664451   28.664451
#> 184   31.200055   31.200055
#> 185   92.061413   92.061413
#> 186  258.922079  258.922079
#> 187  684.652631  684.652631
#> 188  993.198949  993.198949
#> 189  639.481238  639.481238
#> 190  658.302402  658.302402
#> 191 1064.816754 1064.816754
#> 192 1276.720978 1276.720978
#> 193 1360.841680 1360.841680
#> 194 1182.027036 1182.027036
#> 195          NA          NA
#> 196          NA          NA
#> 197          NA          NA
#> 198          NA          NA
#> 199          NA          NA
#> 200          NA          NA
#> 201  123.650701  123.650701
#> 202  145.116192  145.116192
#> 203   48.045989   48.045989
#> 204   60.359199   60.359199
#> 205   81.507819   81.507819
#> 206  108.451321  108.451321
#> 207   63.776428   63.776428
#> 208   20.418998   20.418998
#> 209   33.355506   33.355506
#> 210   88.573683   88.573683
#> 211  298.914826  298.914826
#> 212          NA          NA
#> 213 1398.925758 1398.925758
#> 214          NA          NA
#> 215 1200.994991 1200.994991
#> 216 1145.872629 1145.872629
#> 217  899.925006  899.925006
#> 218  738.948183  738.948183
#> 219 1332.446264 1332.446264
#> 220          NA          NA
#> 221          NA          NA
#> 222          NA          NA
#> 223          NA          NA
#> 224          NA          NA
#> 225          NA          NA
#> 226  148.411096  148.411096
#> 227  277.487816  277.487816
#> 228   97.262213   97.262213
#> 229   90.056298   90.056298
#> 230  139.666138  139.666138
#> 231  102.826129  102.826129
#> 232   18.268087   18.268087
#> 233   44.717650   44.717650
#> 234   91.994244   91.994244
#> 235  579.301178  579.301178
#> 236  829.460542  829.460542
#> 237 1192.706140 1192.706140
#> 238 1302.150322 1302.150322
#> 239 1383.350908 1383.350908
#> 240 1398.860545 1398.860545
#> 241          NA          NA
#> 242          NA          NA
#> 243          NA          NA
#> 244          NA          NA
#> 245          NA          NA
#> 246          NA          NA
#> 247          NA          NA
#> 248          NA          NA
#> 249          NA          NA
#> 250          NA          NA
#> 251  219.828908  219.828908
#> 252  253.159967  253.159967
#> 253  161.398240  161.398240
#> 254  233.493216  233.493216
#> 255  164.022603  164.022603
#> 256   30.894575   30.894575
#> 257   83.088137   83.088137
#> 258  108.862151  108.862151
#> 259  229.800850  229.800850
#> 260  356.712383  356.712383
#> 261          NA          NA
#> 262          NA          NA
#> 263          NA          NA
#> 264          NA          NA
#> 265          NA          NA
#> 266          NA          NA
#> 267          NA          NA
#> 268          NA          NA
#> 269          NA          NA
#> 270          NA          NA
#> 271          NA          NA
#> 272          NA          NA
#> 273          NA          NA
#> 274          NA          NA
#> 275          NA          NA
#> 276  221.718578  221.718578
#> 277  332.692853  332.692853
#> 278  213.463658  213.463658
#> 279  243.120948  243.120948
#> 280   63.730534   63.730534
#> 281  147.329531  147.329531
#> 282  700.109004  700.109004
#> 283  972.574198  972.574198
#> 284  534.244983  534.244983
#> 285  632.502886  632.502886
#> 286          NA          NA
#> 287          NA          NA
#> 288          NA          NA
#> 289          NA          NA
#> 290          NA          NA
#> 291          NA          NA
#> 292          NA          NA
#> 293          NA          NA
#> 294          NA          NA
#> 295          NA          NA
#> 296          NA          NA
#> 297          NA          NA
#> 298          NA          NA
#> 299          NA          NA
#> 300          NA          NA
#> 301   95.553663   95.553663
#> 302  126.688414  126.688414
#> 303   52.805044   52.805044
#> 304   41.873328   41.873328
#> 305    9.222216    9.222216
#> 306  120.766493  120.766493
#> 307  241.896138  241.896138
#> 308  477.623071  477.623071
#> 309  307.530335  307.530335
#> 310  213.983651  213.983651
#> 311          NA          NA
#> 312          NA          NA
#> 313          NA          NA
#> 314          NA          NA
#> 315          NA          NA
#> 316          NA          NA
#> 317          NA          NA
#> 318          NA          NA
#> 319          NA          NA
#> 320          NA          NA
#> 321          NA          NA
#> 322          NA          NA
#> 323          NA          NA
#> 324          NA          NA
#> 325          NA          NA