Overlay SDM predictions onto base geometry

Overlay specified SDM predictions that meet the percent overlap threshold requirement onto base geometry

overlay_sdm(base.geom, sdm, sdm.idx, overlap.perc)

Arguments

base.geom
object of class sfc; base geometry

sdm
object of class sf; original SDM predictions

sdm.idx
names or indices of column(s) with data to be overlaid

overlap.perc
numeric; percent overlap threshold, i.e. percentage of each base geometry polygon must overlap with SDM prediction polygons for overlaid density value to be calculated and not set as NA

Value

Object of class sf with the geometry of base.geom and the data in the sdm.idx columns of sdm overlaid onto that geometry. Note that this means all columns of sdm not in sdm.idx will not be in the returned object. Because the data are considered spatially intensive, the agr attribute will be set as 'constant' for all columns in the returned object

Details

See the eSDM GUI manual for specifics about the overlay process. This process is equivalent to areal interpolation (Goodchild and Lam 1980), where base.geom is the target, sdm is the source, and the data specified by sdm.idx are spatially intensive.

Note that overlay_sdm removes rows in sdm that have NA values in the first column specified in sdm.idx (i.e. sdm.idx[1]), before the overlay. Thus, for valid overlay results, all columns of sdm specified in sdm.idx must either have NA values in the same rows or contain only NAs.

References

Goodchild, M.F. & Lam, N.S.-N. (1980) Areal interpolation: a variant of the traditional spatial problem. Geo-Processing, 1, 297-312.

Examples

overlay_sdm(sf::st_geometry(preds.1), preds.2, 1, 50)

#> Simple feature collection with 325 features and 1 field
#> Attribute-geometry relationship: 1 constant, 0 aggregate, 0 identity
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: -123 ymin: 32 xmax: -117 ymax: 35
#> CRS:            EPSG:4326
#> First 10 features:
#>         Density                       geometry
#> 1  0.0007590157 POLYGON ((-122.875 32.125, ...
#> 2  0.0006972489 POLYGON ((-122.625 32.125, ...
#> 3  0.0010751667 POLYGON ((-122.375 32.125, ...
#> 4  0.0018758023 POLYGON ((-122.125 32.125, ...
#> 5  0.0022005786 POLYGON ((-121.875 32.125, ...
#> 6  0.0023618842 POLYGON ((-121.625 32.125, ...
#> 7  0.0021131219 POLYGON ((-121.375 32.125, ...
#> 8  0.0019932582 POLYGON ((-121.125 32.125, ...
#> 9  0.0020991526 POLYGON ((-120.875 32.125, ...
#> 10 0.0020968369 POLYGON ((-120.625 32.125, ...

overlay_sdm(sf::st_geometry(preds.2), preds.1, c("Density", "Density2"), 50)

#> Warning: The following columns have 'NA' prediction values for different prediction polygons than the first column specified in 'sdm.idx', and thus their overlaid values will be invalid (see the function documentation for more details):
#> Density2

#> Simple feature collection with 1891 features and 2 fields
#> Attribute-geometry relationship: 2 constant, 0 aggregate, 0 identity
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: -123 ymin: 32 xmax: -117 ymax: 35
#> CRS:            EPSG:4326
#> First 10 features:
#>       Density  Density2                       geometry
#> 1  0.18149697 0.2801785 POLYGON ((-122.95 32.05, -1...
#> 2  0.14315669 0.3069032 POLYGON ((-122.85 32.05, -1...
#> 3  0.02813584 0.3870771 POLYGON ((-122.75 32.05, -1...
#> 4  0.02813584 0.3870771 POLYGON ((-122.65 32.05, -1...
#> 5  0.05654899 0.2360303 POLYGON ((-122.55 32.05, -1...
#> 6  0.06602004 0.1856814 POLYGON ((-122.45 32.05, -1...
#> 7  0.05677609 0.3732202 POLYGON ((-122.35 32.05, -1...
#> 8  0.02904423 0.9358365 POLYGON ((-122.25 32.05, -1...
#> 9  0.02904423 0.9358365 POLYGON ((-122.15 32.05, -1...
#> 10 0.03472994 0.2475016 POLYGON ((-122.05 32.05, -1...