Sample SDM density predictions

preds.1, preds.2, and preds.3 are objects of class sf that serve as sample sets of SDM density predictions for the eSDM package

Usage

preds.1

preds.2

preds.3

Format

Objects of class sf with a column of density predictions (name: Density) and a simple feature list column (name: geometry). preds.1 also has a second column of sample density predictions (name: Density2), as well as Var1 and Var2, representing the variance

preds1: An object of class sf (inherits from data.frame) with 325 rows and 5 columns.

preds2: An object of class sf (inherits from data.frame) with 1891 rows and 2 columns.

preds3: An object of class sf (inherits from data.frame) with 1445 rows and 2 columns.

An object of class sf (inherits from data.frame) with 1891 rows and 2 columns.

An object of class sf (inherits from data.frame) with 1445 rows and 2 columns.

Details

preds.1 sample SDM density predictions created by importing Sample_predictions_2.csv into the eSDM GUI, exporting predictions, and then clipping them to the SoCal_bite.csv region. Also manually added two variance columns (numbers are randomly generated with a max of 0.01)

preds.2 sample SDM density predictions created by importing Sample_predictions_1.csv into the eSDM GUI, exporting predictions, and then clipping them to the SoCal_bite.csv region

preds.3 is a set of sample SDM density predictions created by importing Sample_predictions_4_gdb into the eSDM GUI, exporting predictions, and then clipping them to the SoCal_bite.csv region